test add job error handling
We want the interface to a test to appears as near as possible to be
a direct python object load and execute.  We are forking for safety
and recoverability, but we don't want to have to think about that.
Add an exception handler to the child to pickle the exception and
then load that in in the parent and raise it on.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@17 592f7852-d20e-0410-864c-8624ca9c26a4