errors: fix sub process exception handling
Ensure that we pick up the full stack from any test errors from Python.
We now classify any failed error as an UnhandledError.  This allows
us to store the stack trace in that error and emit it more cleanly
into the logs.

Also adds an unused non-fork version if we decided to remove the
fork here.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@37 592f7852-d20e-0410-864c-8624ca9c26a4