job execute script in separate namespace
We currently execute the users job in the top level namespace
for autotest.  This does two things: it pollutes the test namespace
with varying autotest information which encourages its use; and
it forces us to execute the script from the file autotest.  Move the
execution into a virgin namespace (with just error and autotest_utils
imported), and move execution into the job object where it belongs.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@112 592f7852-d20e-0410-864c-8624ca9c26a4