Add xmtest Xen test
This adds the basic guts to get xmtest running under a Xen domain 0.  Xmtest 
is currently integrated in the Xen source tree, but rather than needing to 
copy the whole thing here just for the tests, I would recommend giving it a 
tarball with just the xm-test directory from the xen-unstable hg tree, and 
that's what it's expecting right now. Instructions with where to get it are 
in the patch.

Signed-off-by: Paul Larson <pl@us.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@186 592f7852-d20e-0410-864c-8624ca9c26a4