self-test: move to using /bin/pwd to cover for symlinked directories
When we are parsing the output of tests we want to remove
the directory name from the output so that test results are
not location specific.  However 'pwd' lies in bash, and worse
even 'command pwd' which is supposed to avoid the builtin
really runs the builtin!  Move to using /bin/pwd to get the
truth.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@200 592f7852-d20e-0410-864c-8624ca9c26a4