fix output sequencing and location for tests
When a test completes we pop the fdstacks and drop back to the
job level logs.  However if we take an exception including any job
failure then we miss this unstack.  This leads to the job output
being logged incorrectly and test output to be miss-ordered.
Ensure we unstack the test fds at the end of the job.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@212 592f7852-d20e-0410-864c-8624ca9c26a4