tests: add downloadable test support v3
[Ok, after much feedback from Martin and generally cleaning things
up here is the latest version of the downloadable test support.]

Add support for downloading and installing tests.  Tests are loaded
from a URL and stored locally in a per-host hierachy.  This patch
extends the job.runtest interface such that it can optionally accept
a testname (for builtin tests) or a URL for external tests:

    job.runtest('out', 'kernbench', args)
    job.runtest('out', 'http://www.somewhere.com/test/test.tgz', args)

An external test is downloaded into a tests/download/<group>/<test>
and executed there.  Group is derived from the URL which allows
the test to both be separated in the filesystem namespace, and in
the python namespace during execution.  This allows two different
mytest's to exist at different URLs without a clash.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@288 592f7852-d20e-0410-864c-8624ca9c26a4