This patch further updates the dbt2 test to execute against pgpool.
This patch further updates the dbt2 test to execute against pgpool.  It
also supplies a config file for pgpool to run in connection pooling
mode.  The dbt2 control file has been updated with test notes (via -z
flag) as opposed to hard coding them in the dbt2.py script, which was
done previously.

Signed-off-by: Mark Wong <markw@osdl.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@294 592f7852-d20e-0410-864c-8624ca9c26a4