hook kernel class to add a newly built kernel to the bootloader.
I've been playing around with adding entries to the bootloader with and
without mkintrds and the following patch adds a simple method to the
kernel class for adding a newly built kernel to the bootloader.  I've
also included a new sample control file which exercises the new method.
This patch relies on patch 1 from this set to ensure the mkinitrd is
added to the bootloader entry.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@320 592f7852-d20e-0410-864c-8624ca9c26a4