kernel: add a direct boot() method
In my mind I want to build, then boot a kernel I have made.  I do not
want to think about bootloaders, or boot entries or any of that.  Just
want to make a kernel out of my tarball and make this machine start
running it.

Add a boot() method to the kernel object which installs the kernel
as autotest and boots it with the passed arguments.

	kernel.build()
	kernel.boot()

Signed-off-by: Andy Whitcroft <apw@shadowen.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@363 592f7852-d20e-0410-864c-8624ca9c26a4