harness: ensure the configuration is initialised
The harness is very likely to want to modify the job configuration.
It needs a chance to update the configuration before anyone uses it.
We could add a new 'config_ready' style hook, but the configuration
is entirly passive at this time so we can simply initialise it
before the harness.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@400 592f7852-d20e-0410-864c-8624ca9c26a4