Fix Bonnie results format.
From: Ryan Stutsman <stutsman@google.com>

This patch is in respond to a concern raised by Kamal Mansouri about
the bonnie test keyval results format.

Essentially it bonnie reports size in M and G and this expands either
to bytes in the keyval file.  Also, invalid results were reported as
strings of + in the keyval file.  These are now replaced with 0.



git-svn-id: http://test.kernel.org/svn/autotest/trunk@558 592f7852-d20e-0410-864c-8624ca9c26a4