This patch increases the timeout for the start barrier.
It used to be 30 seconds. When executing the netperf test on two machines of different specs, the most powerful one would have the time to finish compiling and timeout on the barrier while the other was still compiling. 600 seconds should leave ample tolerance at the cost of a longer wait if there is indeed something that goes wrong. 

From: Benjamin Poirier <poirier@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@612 592f7852-d20e-0410-864c-8624ca9c26a4