harness: add a status with detail hook
We really could do with getting the status in its raw components to
better generate the required format for the server harness.  Add a
test_status_detail() hook which passes the code, subdir, operation, and
message separatly.  Use this hook within the ABAT harness to reform the
intermediate status messages into acceptable format.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@687 592f7852-d20e-0410-864c-8624ca9c26a4