Update mirror/trigger engines.
From: Ryan Stutsman <stutsman@google.com>

Ok, final rework to this mirror script.  From here bugs should be easy
to fix as patches.

This version should address all the concerns you outlined in your
prior looks at the mirror script.

The major addition of this version is automatic smart kernel config
selection for builds.

It works like this:  you pass a directory to mirror via the -f flag
like so:

mirror -f /home/testing/config -c /home/testing/client_test_mapping
/home/testing/mirror /home/testing/queue

Now, when a new kernel release is found it will test that kernel on
each machine listed in the client_test_mapping file with each of the
control files listed on that line.

For example, if client_test_mapping contains:

machine1 ~/basic-tests ~/do-something.srv
machine2 ~/basic-tests

It will queue the AutoTest control file ~/basic-tests on both machines
and the ~/do-something.srv Autoserv file just on machine1.  The
Autoserv file support should make it quite easy to do similar testing
with RPMs, for example, though the mirroring urls and regexps would
probably have to modified or configurable.

In doing this for AutoTest control files it will wrap it with the code
necessary to build and boot the kernel version it mirrored.  If there
is a file in /home/testing/config with the name machine1, for example,
it will use that to determine which kernel build config should used.

If there where a /home/testing/config/machine1 like this:

2.6.0 /boot/config-2.6.18
2.6.19 /boot/config-2.6.21

Then for any kernel 2.6.0 or later it will build it with
/boot/config-2.6.18 (path on the *client* machine).  For any kernel
2.6.19 or later it will use /boot/config-2.6.21.
If no file exists for that machine it will simply use /boot/config.
Also, if the kernel version to be tested it older than any config in
the mapping it will automatically use the oldest in the file (2.6.0 ==
/boot/config-2.6.18) in this case.

The /home/testing/mirror arg tells it to mirror the kernel.org
releases to that directory and /home/testing/queue contains the queues
that correspond to the names given in client_test_mapping.

Hopefully this all makes sense.  I'll be watching the list to see if
there are questions or bugs to work out with the script.  I have
tested it somewhat and it seems to work as expected.




git-svn-id: http://test.kernel.org/svn/autotest/trunk@692 592f7852-d20e-0410-864c-8624ca9c26a4