Fix for autoserv's param handling
From: jorlow@google.com

OptionParser (and getopt) assume that they know of all options your program
uses.  Unfortunately, this means that control files that want to have their own
options are out of luck.  This patch parses the machine option manually in order
to get around this.  It also now passes extra args to the control file (like
what was done in a much older version).




git-svn-id: http://test.kernel.org/svn/autotest/trunk@700 592f7852-d20e-0410-864c-8624ca9c26a4