ABAT: handle even obscurer SLES root specifier
It seems that in SLES images root will be shown mounted on /dev/root
but that /dev/root will not even exist.  At this point we have use
the mtab to determine where / is mounted from.  Use df -lP to find
the underlying device for passing to the boot loader.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@734 592f7852-d20e-0410-864c-8624ca9c26a4