Fixing bonnie results parser
When running bonnie++ using the '-s' argument defining a chunk
size with a file size, the 'format_results' method gives an exception
when trying to convert the file size argument.

E.g:
job.run_test('bonnie', extra_args = '-s 8212:512')

Exception:
ValueError: invalid literal for int(): 8212M:512

Looking at the bonnie man page, we have the following:
[-s size(Mb)[:chunk-size(b)]]

-s     the   size  of  the  file(s)  for  IO  performance  measures  in
       megabytes. If the size is greater than 1G  then  multiple  files
       will  be  used to store the data, and each file will be up to 1G
       in size.  This parameter may include the  chunk  size  seperated
       from  the  size by a colon.  The chunk-size is measured in bytes
       and must be a power of two from 256  to  1048576.   NB  You  can
       specify  the  size in giga-bytes or the chunk-size in kilo-bytes
       if you add 'g' or 'k' to the end of the number respectively.

So, basically we can run bonnie giving a specific chunk size (in bytes,
or in kilo-bytes if you give the number followed by 'k').

As we can define the chunk size, I also added it at keyval results
(with 0 when the user doesn't define it).

This patch also fix the problem of parsing the results when we have more
than one iteration (it was getting only the last one).

Signed-off-by: Ricardo Salveti de Araujo <rsalvetidev@gmail.com>




git-svn-id: http://test.kernel.org/svn/autotest/trunk@751 592f7852-d20e-0410-864c-8624ca9c26a4