boottool: add a common method to build the new kernel cmdline v2
We have two different implementations of the kernel cmdline editor; the
code which implements --args= --remove-args=.  One of these implement
addition as replace where the argument already exists, the other does not.

From the code it appears that they are really intended to work on
independant keyword=value sets, and a single one at a time.  However,
this means that to remove the current args and add a set of new ones
you may have to perform a large number of removes followed by adds,
each rewriting the config.

This patch pulls out this processing to _build_args on the base class.
This new implementation treats the arguments to --args= and --remove-args=
as multiple keyword=value pairs and applies all of them, removes first
then additions.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@779 592f7852-d20e-0410-864c-8624ca9c26a4