Integrate building from git repo into kvm class
Add support for the kvm class to be built from the kvm git repo.  After pulling
down the latest kvm and kvm-userspace git repositories, use kvm-userspace's make
sync method for pulling in the latest kvm kernel bits.

Add a dict to track full path to modules for running with modinfo to extract kvm
version for scripts.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1033 592f7852-d20e-0410-864c-8624ca9c26a4