boottool: take all default values from the first valid entry
We currently find defaults for kernel bootpath, arguments, and target
partitions from the default entry.  However this may not be a full kernel
entry, for example in grub it may be a chainload entry.  So if the default
is not a kernel entry, search all for one which carries a kernel and use
that.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@1095 592f7852-d20e-0410-864c-8624ca9c26a4