Revert broken exception handling change
Can we revert this part of the change? We were not raising a string exception 
here. pickle.load() deserializes the exception raised on the client side. This
change effectively converts all exceptions thrown at the client into an 
AutotestError.

For me, this breaks the case where a TestError is raised. job.run_test()
won't re-raise a TestError:

		if exc_info and isinstance(exc_info[1], TestError):
			return False
		elif exc_info:
			raise exc_info[0], exc_info[1], exc_info[2]
		else:
			return True

I have control files that call the same test many times with different
parameters used for each iteration. If there is a test error for one iteration,
the next test is run immediately. After this change, the entire control
file gets aborted after the first TestError gets raised. I reverted
the change in my local copy and confirm that the original behavior
returns.

Below is a patch to revert the change.

Signed-off-by: Mandeep Singh Baines <msb@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1174 592f7852-d20e-0410-864c-8624ca9c26a4