Modifying dacapo test - remove need for configuration files
With the extended configuration parser, it's better to keep all the
dacapo test information inside the control file. Of course, the control
file is only a possible example of usage. It's also possible to load
test configuration from a local or a remote file, as it is shown on the
config file for the test.

Signed-off-by: Lucas Meneghel Rodrigues <lucasmr@br.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1213 592f7852-d20e-0410-864c-8624ca9c26a4