Fixing debian support for LSB testing
This patch fixes the final bug (see message [PATCH 1/2] Fixing debian
support for LSB testing): when we decided to install a lsb manager
different from the one that comes with the lsb bundle (tarball), I
forgot to consider the debian case and made the test try to install the
new package straight away.

lsb_dtk.py fixes:
 * If it's a debian based distro, convert the lsb_dtk manager rpm to
debian before trying to install it.

Signed-off-by: Lucas Meneghel Rodrigues <lucasmr@br.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1393 592f7852-d20e-0410-864c-8624ca9c26a4