Make package.convert() on package.py to actually return the absolute path for the converted package.
Description
Even though the LSB debian support was fixed, the package.convert
function was still not returning the absolute path of the converted
package. This patch fixes this mistake. Also, makes the convertion
function to print the operation summary (Package xxx.rpm was
successfully converted to xxx.deb or the similar)

Signed-off-by: Lucas Meneghel Rodrigues <lucasmr@br.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1394 592f7852-d20e-0410-864c-8624ca9c26a4