Make the step engine more powerful and add a soft-reboot test that uses it
This patch changes the usage of job.next_step while maintaining backwards compatibility.  You can now pass named parameters and you no longer need to wrap everything in an array.  What used to be job.next_step(['some_function', 'some_param', 'another param']) can now be job.next_step('some_function', 'some_param', 'another_param'). Like before, the function passed in can be the actual function or a string representation of that function.  Also like before, the function must be in the global namespace fo the control file.  For backward compatibility, if the first parameter is a list, the first item in the list is assumed to be the function and the rest are assumed to be arguments.  In other words, this change should not break anything and should make the next step syntax cleaner.

From: jorlow@google.com



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1465 592f7852-d20e-0410-864c-8624ca9c26a4