-include acl-inaccessible hosts in ineligible_host_queues blocks. This should make metahosts obey ACLs. -recompute blocks for all non-complete jobs after any acl changes -make the scheduler clear out blocks for completed jobs
This is messier than I would've liked because it required two things
Django is not very good at - subqueries and notification of every time
a many-to-many relationship changes.  I used a custom manager and a
custom manipulator to get around the two, respectively, in as clean a
way I could find.

Signed-off-by: Steve Howard <showard@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1471 592f7852-d20e-0410-864c-8624ca9c26a4