Risk: Medium Visibility: Should remove a case where "real" errors get masked by an accidental double-call of parser.end().
This change changes the parser.end() call to explicitly catch a
StopIteration error and keep going with a warning, rather than just
letting the exception propagate up. This can only happen if end() is
called multiple times, which is an error but not a serious enough one
to warrant a process-ending exception.

Signed-off-by: John Admanski <jadmanski@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1514 592f7852-d20e-0410-864c-8624ca9c26a4