Replaced "from common_lib.utils import *" with "from common_lib import utils" and added the necessary methods which just call the methods in common_lib.utils.
From our conversation, John and I have looked at client.common_lib.utils and server.utils and have sketched out ideas for how much of this will be refactored as we move forward.  This patch below is just a temporary place holder to keep things working and make them testable so that as we re-factor we can have confidence that things are still working and regression free.

From: Travis Miller



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1568 592f7852-d20e-0410-864c-8624ca9c26a4