Allow subtitle above plot's key legends.
When X dimension values and the plotted dataset lines have similar names, a subtitle
is needed just above the key legends, to help clarify how they are different from
the X axis labels.  In this same situation, it is helpful to position the
key legend outside and to the right of the graph, rather than immediately below
the X axis labels.

Signed-off-by: Duane Sand <duanes@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1654 592f7852-d20e-0410-864c-8624ca9c26a4