Summary: Fixing circular reference issue during barrier module import.
The issue was breaking nearly all autotest modules, since it's right on
the common_lib setup. The fix was simple, just remove the circular
reference to common. Also, changed the placement of the __author__
decorator, and transferred the BarrierError class to the error module,
for the sake of consistency (all other modules define their exceptions
on error). This patch is the fix for the bug #4
(http://test.kernel.org/trac-autotest/ticket/4)

Visibility: High (Makes all modules that import common to work again)

Signed-off-by: Lucas Meneghel Rodrigues <lucasmr@br.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1840 592f7852-d20e-0410-864c-8624ca9c26a4