Fix the parser behaviour when handling non-grouped tests inside of a client job. The test subdir should override the job subdir (or rather, the job subdir should not override the test subdir).
The main idea of this fix is to move the code that check if the line
currently being parsed is an unexpected abort to before the line is
actual parsed and the "current" test state of the parser is updated.
This allows us to abort the current test.

Risk: Medium
Visibility: An unexpected client failure during a test should look
            like a job abort, not a build+job abort.

Signed-off-by: John Admanski <jadmanski@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@1923 592f7852-d20e-0410-864c-8624ca9c26a4