- Add a set_autodir method to ssh_host to set the autotest installation directory value - Set the autotest installation directory in ssh_host from server/autotest.py once the installation of autotest is completed on that machine.
The idea behind the patch is that host.get_autodir is generally not
very useful, because hosts are generally constructed with
autodir=None. But for server-side code outside of autotest.py, there
really isn't any other way to figure out where autotest is installed
on the remote machine.

So this patch changes autotest.py to change the autodir on a host
object when it uses that host object to install autotest. That way
code using the host object can figure out where autotest is, as long
as it's been installed.

Signed-off-by: Ashwin Ganti <aganti@google.com>




git-svn-id: http://test.kernel.org/svn/autotest/trunk@2016 592f7852-d20e-0410-864c-8624ca9c26a4