Provide absolute path to init on client/bin/harness_standalone.py
Description: On distros that have an /etc/event.d directory, if we
execute autotest with a non-root user, /sbin usually won't be on the
PATH environment variable, so autotest won't be able to find init hence
throwing a CmdError. We can fix that by just adding /sbin to the command
line. Since init is under /sbin on pretty much all linux distros, it's a
fairly safe change to make.

Risk: Low - See description
Visibility: Low - Only users executing autotest with a non-root user.

Signed-off-by: Lucas Meneghel Rodrigues <lucasmr@br.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2081 592f7852-d20e-0410-864c-8624ca9c26a4