- add a server.hosts.base_classes.Host.verify() abstract method  - add a server.hosts.base_classes.Host.check_diskspace(path, gb) helper method  - implement verify() in server.hosts.ssh_host.SSHHost, from server.control_segments.verify.verify() function
The point is that the actual verify code is composed of a not so generic
action (check that available diskspace is >20GB) then, if any, run a
site_verify function. This is not as extensible and generic as subclassing
methods  and I was thinking that Host class (and its derivatives) is the
right place to implement machine verification.

The same for repair.

From: Jean Parpaillon <jean.parpaillon@kerlabs.com>




git-svn-id: http://test.kernel.org/svn/autotest/trunk@2211 592f7852-d20e-0410-864c-8624ca9c26a4