Add server-side sysinfo collection code (for server tests). This is implemented by doing a temporary lightweight (no tests or deps) autotest install on the remote machine and then running a small script to generate the sysinfo.
The code is somewhat ugly because of the nature of what it has
to do on the remote machine (it needs to run before the test,
save some state, and then run afterwards with the saved state).
I accomplish this by pickling off of the sysinfo object after
the pre-test run, and reloading it during the post-test run.
However, I don't see a cleaner solution; we can't depend on
anything we leave behind on the remote machine while the test
is running not getting smashed.

This only applies to single-machine jobs (or multi-machine jobs
that break down into single-machine jobs via job.parallel_simple).
For multi-machine jobs we'd need to come up with a new,
multi-machine way to log sysinfo; in my mind it's a completely
separate problem.

Risk: Medium
Visibility: Adds sysinfo collection to server-side tests.

Signed-off-by: John Admanski <jadmanski@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2270 592f7852-d20e-0410-864c-8624ca9c26a4